feat: 룸메찾기 화면 수정, 추가조건 설정 UI 추가

+++ CBNU-domitory/lib/models/similarity.dart
@@ -1,5 +1,5 @@
 class Similarity {
-  final int similarity;
+  final double similarity;
   final List<String> similar_factors;
 
   Similarity({required this.similarity, required this.similar_factors});
